requirements.txt
streamlit
agno
groq
psycopg2-binary
python-dotenv
dotenv
PyPDF2
pandas
yfinance
sqlalchemy